// src/lib/api/generated/models/updateClientPayload.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Marketplace API
 * API for Village Website & Marketplace
 * OpenAPI spec version: 1.0.0
 */

export interface UpdateClientPayload {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  city?: string;
  state?: string;
  country?: string;
  zip?: string;
}